For #756 - move SampleFilter out of db
